# backend/src/main/resources/openapi/sims-openapi.yaml
openapi: 3.0.3
info:
  title: SIMS API — Exam, Fee and Communication modules
  description: |
    School Information Management System — the exam, fee and communication modules
    after the business-workflow redesign.

    ## Conventions

    **Nested business objects.** Responses embed the data a screen needs rather than
    bare foreign keys: `subject`, `student`, `schoolClass`, `createdBy` and
    `updatedBy` are objects carrying ids *and* names. The old flat `*Id` fields are
    still returned but are deprecated and will be removed once the React app has
    migrated off them.

    **Pagination.** List endpoints return a Spring `Page`:
    `?page=0&size=20&sort=createdAt,desc`.

    **Actor fields.** Requests carry `actorId` (admin staff) or `enteredById` /
    `updatedById` (teacher) because authentication is still being built by another
    developer. These fields disappear once the JWT principal is available.

    **Marks and fee "sheets".** Marks and fee statuses are worked on a section at a
    time, so each has a `GET /sheet` that returns the whole roster pre-filled and a
    `PUT /sheet` that saves it in one transaction.
  version: 2.0.0
  contact:
    name: SIMS Team
servers:
  - url: http://localhost:8081
    description: Local development
  - url: https://api.example.com
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/fees/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Fees]
      summary: Get fee by id
      operationId: getFeeById
      responses:
        '200':
          description: Fee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeeResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Fees]
      summary: Update a fee
      description: |
        Changing the class or academic year is rejected once student statuses exist,
        since it would move those students' records onto a different class or year.
      operationId: updateFee
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeeRequest' }
      responses:
        '200':
          description: Fee updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeeResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Duplicate definition, or statuses already recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Fees]
      summary: Delete a fee
      description: Rejected while any student fee status references it.
      operationId: deleteFee
      responses:
        '204': { description: Fee deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Student fee statuses are recorded against this fee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Student Fee Status
  # ==================================================================
  /api/v1/student-fee-status/sheet:
    get:
      tags: [Student Fee Status]
      summary: Fee collection sheet
      description: |
        One fee, one section, every student's payment status plus counts by status.
        `status` is `null` for students with no record yet.
      operationId: getFeeStatusSheet
      parameters:
        - name: feeId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: sectionId
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Fee status sheet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeeStatusSheetResponse' }
        '400':
          description: The section does not belong to the fee's class
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Student Fee Status]
      summary: Save a fee collection sheet
      description: |
        Transactional bulk upsert, validated in full before anything is written.
        Exactly one of `updatedById` (teacher) or `updatedByAdminId` must be given.
      operationId: saveFeeStatusSheet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FeeStatusSheetSaveRequest' }
      responses:
        '200':
          description: Sheet saved; the refreshed sheet is returned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeeStatusSheetResponse' }
        '400':
          description: |
            Validation failure, both or neither actor supplied, a duplicate enrollment
            in the payload, or an enrollment outside the section.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/student-fee-status/defaulters:
    get:
      tags: [Student Fee Status]
      summary: Students who have not fully paid
      description: Statuses of `NOT_PAID` or `PARTIAL`, optionally narrowed by class and year.
      operationId: getFeeDefaulters
      parameters:
        - $ref: '#/components/parameters/ClassIdFilter'
        - $ref: '#/components/parameters/AcademicYearIdFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Page of outstanding statuses
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageStudentFeeStatusResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/student-fee-status:
    get:
      tags: [Student Fee Status]
      summary: List student fee statuses
      operationId: getAllStudentFeeStatuses
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Page of statuses
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageStudentFeeStatusResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    post:
      tags: [Student Fee Status]
      summary: Record a student's fee status
      description: |
        One record per (enrollment, fee). The student must be enrolled in the class
        the fee belongs to.
      operationId: createStudentFeeStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentFeeStatusRequest' }
      responses:
        '200':
          description: Status recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentFeeStatusResponse' }
        '400':
          description: |
            Validation failure, both or neither actor supplied, or the student is not
            in the fee's class.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: A status already exists for this student and fee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/student-fee-status/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [Student Fee Status]
      summary: Get status by id
      operationId: getStudentFeeStatusById
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentFeeStatusResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    put:
      tags: [Student Fee Status]
      summary: Change a student's fee status
      operationId: updateStudentFeeStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StudentFeeStatusRequest' }
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StudentFeeStatusResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: A status already exists for the target student and fee
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

    delete:
      tags: [Student Fee Status]
      summary: Delete a status record
      operationId: deleteStudentFeeStatus
      responses:
        '204': { description: Status deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/student-fee-status/enrollment/{enrollmentId}:
    parameters:
      - name: enrollmentId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Student Fee Status]
      summary: All fee statuses for a student
      description: Bounded by the number of terms, so returned unpaginated.
      operationId: getStudentFeeStatusesByEnrollment
      responses:
        '200':
          description: Statuses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/StudentFeeStatusResponse' }
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ==================================================================
  # Announcements
  # ==================================================================
...
components:
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, default: 20 }
    Sort:
      name: sort
      in: query
      required: false
      description: '`property,(asc|desc)`; repeatable.'
      schema:
        type: array
        items: { type: string }
      example: ['createdAt,desc']
    ClassIdFilter:
      name: classId
      in: query
      required: false
      schema: { type: string, format: uuid }
    AcademicYearIdFilter:
      name: academicYearId
      in: query
      required: false
      schema: { type: string, format: uuid }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the login endpoint. Authentication is still being built, so
        endpoints are currently open and role rules are recorded as TODOs beside each
        controller method.
  responses:
...
    NotFound:
      description: A referenced record does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 404
            error: Not Found
            message: 'Exam not found with id: 9f1c8b2e-...'
            path: /api/v1/exams/9f1c8b2e-...
    ValidationError:
      description: Bean validation failure; offending fields are listed in `fieldErrors`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            timestamp: '2026-08-02T10:30:00Z'
            status: 400
            error: Bad Request
            message: Validation failed
            path: /api/v1/exams
            fieldErrors:
              examName: Exam name is required
              actorId: Acting admin id is required
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    FeeStatus:
      type: string
      enum: [PAID, PARTIAL, NOT_PAID]
...
    ActorSummary:
      type: object
      description: A person who created or last changed a record.
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: R. Iyer }
        role:
          type: string
          description: Designation, e.g. PRINCIPAL or SENIOR_TEACHER.
          example: SENIOR_TEACHER
    ClassSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Grade 10 }
...
    AcademicYearSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string, example: 2025-2026 }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
    SectionSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: A }
        classId: { type: string, format: uuid }
        className: { type: string, example: Grade 10 }
        displayName:
          type: string
          description: Ready-made heading label.
          example: Grade 10 - A
    StudentSummary:
      type: object
      description: A student seen through their enrollment.
      properties:
        enrollmentId: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        name: { type: string, example: Aarav Sharma }
        rollNumber: { type: string, example: '04' }
        admissionNumber: { type: string, example: ADM-2025-0142 }

    # ---------------- errors & paging ----------------
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 409 }
        error: { type: string, example: Conflict }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: { type: string }
          description: Present only for bean validation failures.
      required: [timestamp, status, error, message, path]
    PageMetadata:
      type: object
      properties:
        totalElements: { type: integer, format: int64, example: 137 }
        totalPages: { type: integer, example: 7 }
        number: { type: integer, description: Zero-based page index, example: 0 }
        size: { type: integer, example: 20 }
        numberOfElements: { type: integer, example: 20 }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }

    # ---------------- exam ----------------
...
    FeeSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        termName: { type: string, example: Term 1 }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        academicYearLabel: { type: string, example: 2025-2026 }
    FeeRequest:
      type: object
      required: [classId, academicYearId, termName, actorId]
      properties:
        classId: { type: string, format: uuid }
        academicYearId: { type: string, format: uuid }
        termName: { type: string, maxLength: 255, example: Term 1 }
        actorId: { type: string, format: uuid }
    FeeResponse:
      type: object
      properties:
        feeId: { type: string, format: uuid }
        termName: { type: string }
        schoolClass: { $ref: '#/components/schemas/ClassSummary' }
        academicYear: { $ref: '#/components/schemas/AcademicYearSummary' }
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        classId: { type: string, format: uuid, deprecated: true }
        academicYearId: { type: string, format: uuid, deprecated: true }
    StudentFeeStatusRequest:
      type: object
      required: [enrollmentId, feeId, status]
      description: Exactly one of `updatedById` or `updatedByAdminId` must be supplied.
      properties:
        enrollmentId: { type: string, format: uuid }
        feeId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/FeeStatus' }
        updatedById:
          type: string
          format: uuid
          nullable: true
          description: Teacher changing the status.
        updatedByAdminId:
          type: string
          format: uuid
          nullable: true
          description: Admin staff changing the status.
    StudentFeeStatusResponse:
      type: object
      properties:
        studentFeeStatusId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/FeeStatus' }
        student: { $ref: '#/components/schemas/StudentSummary' }
        section: { $ref: '#/components/schemas/SectionSummary' }
        fee: { $ref: '#/components/schemas/FeeSummary' }
        updatedBy:
          allOf: [{ $ref: '#/components/schemas/ActorSummary' }]
          description: Whoever last changed the status — teacher or admin, collapsed into one field.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        enrollmentId: { type: string, format: uuid, deprecated: true }
        feeId: { type: string, format: uuid, deprecated: true }
        updatedById: { type: string, format: uuid, deprecated: true }
    FeeStatusSheetEntry:
      type: object
      properties:
        student: { $ref: '#/components/schemas/StudentSummary' }
        studentFeeStatusId: { type: string, format: uuid, nullable: true }
        status:
          allOf: [{ $ref: '#/components/schemas/FeeStatus' }]
          nullable: true
          description: Null when no status has been recorded yet.
        updatedBy: { $ref: '#/components/schemas/ActorSummary' }
        updatedAt: { type: string, format: date-time, nullable: true }
    FeeStatusSheetResponse:
      type: object
      properties:
        fee: { $ref: '#/components/schemas/FeeSummary' }
        section: { $ref: '#/components/schemas/SectionSummary' }
        totalStudents: { type: integer, example: 40 }
        paidCount: { type: integer, format: int64, example: 31 }
        partialCount: { type: integer, format: int64, example: 4 }
        notPaidCount: { type: integer, format: int64, example: 3 }
        unrecordedCount:
          type: integer
          format: int64
          description: Students with no status record yet.
          example: 2
        entries:
          type: array
          items: { $ref: '#/components/schemas/FeeStatusSheetEntry' }
    FeeStatusSheetEntryRequest:
      type: object
      required: [enrollmentId, status]
      properties:
        enrollmentId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/FeeStatus' }
    FeeStatusSheetSaveRequest:
      type: object
      required: [feeId, sectionId, entries]
      description: Exactly one of `updatedById` or `updatedByAdminId` must be supplied.
      properties:
        feeId: { type: string, format: uuid }
        sectionId: { type: string, format: uuid }
        updatedById: { type: string, format: uuid, nullable: true }
        updatedByAdminId: { type: string, format: uuid, nullable: true }
        entries:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/FeeStatusSheetEntryRequest' }

    # ---------------- communication ----------------
...
    PageStudentFeeStatusResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/StudentFeeStatusResponse' }
...